P5/requirements.txt
pymongo==4.6.1
pandas==2.1.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
werkzeug==3.0.1
kagglehub>=0.2.0

